// src/helpers/types.ts
/* API TYPES */

type Info = {
	count: number;
	pages: number;
	next: string | null;
	prev: string | null;
};

type Location = {
	name: string;
	url: string;
};

type Origin = {
	name: string;
	url: string;
};

export type CharacterObj = {
	id: number;
	name: string;
	status: string;
	species: string;
	type: string;
	gender: string;
	origin: Origin;
	location: Location;
	image: string;
	episode: string[];
	url: string;
	created: string;
};

export type CharactersResponse = {
	info: Info;
	results: CharacterObj[];
};